Disable juint reports tmp
The junit reports are creating 90mb files inside of redis for
`gitlab-ce/ee` project. There is an ongoing production incident
https://gitlab.com/gitlab-com/gl-infra/production/issues/928 that is
because of high CPU uitlaization on the redis nodes. GitLab-ce/ee are
the highest consumers of memory becuase of these junit reports.

This is a bandage until we can scale redis properly.