Fix subgroup url in search drop down
Inside a subgroup scope, when the user clicks on the
search drop down in the header, some options are displayed.

Some of this links direct the user to the issues and merge
requests linked to the subgroup. Nevertheless, the url
was broken because they doesn't include the complete
group namespace