Bring buttons style up to design spec
This is a CSS effort only.

Fixes the padding of the buttons to be `8px 12px`
(including border) so that the button height is
32px.

Also adjusts the border width for all buttons when
the state is `hover`, `focus`, or `active to be 2px
thick instead of 1px thick. This is accomplished
through an inset box-shadow as not to increase
the size of the button.

Fixes some colors for primary and tertiary buttons.