Remove duplicated mapping key in config/locales/en.yml
The "errors" section was defined twice in this file, the first one was
the once being removed here, another one is at line 164 (before removal)
or line 158 (after removal).

The content of duplicated strings definition are totally the same, the
second part has more data (41 lines in total, compare to the removed 6
lines), so I removed the first part but not the second part.

cc !16710