Build correct variables for testing translations
This makes sure we build the correct variables for testing
translations.

When translating, we could be specifying the variables in different
forms for each id:

- In the singular we could be using a `%{hash}` interpolation
- In the plural we could be using a `%d` interpolation

This changes the tests to accommodate for that: We now use the
variables used in the relevant translation id as the source for the
variables we mix in in specs.