Refactor protected branch resource and spec
Rename Resource::Branch to Resource::ProtectedBranch to be clear that
it configures the settings to protect a branch, rather than just
creating a new branch.

Make branch an attribute to make it clear creating the branch is not
the main responsibility of the class.

Initialize the project with a readme so that the resource only has
to push once when it creates a branch.

Stop ProjectPush from visiting the project page so that it doesn't
have to use the browser UI.