Fix global state in registrations_controller_spec
This spec would mess with global state used for captchas, but not reset
the state in all cases. This would then lead to the spec
spec/controllers/snippets_controller_spec.rb failing, but only on a fork
of gitlab-ee. To fix this we ensure the state is properly reset after
running the specs that change this state.